refactor(cli): route all gateway calls through the shared gateway-client
- soul.ts and claw.ts drop their private api() fetch wrappers and use
  apiFetch() — they now attach the Authorization header and unwrap the
  gateway error envelope like every other subcommand (previously they
  silently 401'd against an --auth-protected gateway)
- the nonstandard OWNPILOT_API_URL env var is removed; both commands
  honor the standard OWNPILOT_GATEWAY_URL; clawWatch derives its WS URL
  from the shared base via the new getBaseUrl() export
- ensureGatewayError() is exported from gateway-client.ts; the three
  identical local copies in skill.ts, channel.ts, tunnel-wizard.ts are
  deleted
- error-path test helpers updated to the shared client's semantics
  (json error envelope instead of text())
