+++ packages/cli/src/commands/skill.ts
@@ -49,18 +49,7 @@ interface PermissionInfo {
 // ============================================================================
 // apiFetch + auth-header attachment lives in `./gateway-client.ts`.
 
-import { apiFetch, gatewayUnreachableMessage } from './gateway-client.js';
-
-function ensureGatewayError(error: unknown): never {
-  const hint = gatewayUnreachableMessage(error);
-  if (hint) {
-    console.error(hint);
-  } else {
-    const msg = error instanceof Error ? error.message : String(error);
-    console.error(`\nError: ${msg}\n`);
-  }
-  process.exit(1);
-}
+import { apiFetch, ensureGatewayError } from './gateway-client.js';
 
 async function listExtensions(): Promise<ExtensionInfo[]> {
   const data = await apiFetch<ExtensionInfo[] | ExtensionsListResponse>('/extensions');
